Unify event handling into single event handler lambda
